Refactor train_model function and modify code cells in analitycs notebook
The 'train_model' function in training.py is refactored to improve the train-test split. Labels were isolated in a separate variable from the dataframe before the splitting process for clarity. The splitting ratios are also adjusted and a random state is set for reproducibility. Code cells in analitycs.ipynb were modified for better data inspection and presentation. Commentary was added for clarity and the data information checkout processes were extended.